+++ lib/Assyst.js
@@ -1,4 +1,4 @@
-const config = require('./config.json');
+const config = require('../config.json');
 const eris = require('eris');
 const fs = require('fs')
 const util = require('util')
