+++ lib/Command.js
@@ -1,8 +1,8 @@
 const config = require('../config.json')
-const Utils = require('./utils.js').Utils
+const Utils = require('../src/utils.js').Utils
 const utils = new Utils
 const superagent = require('superagent')
-const assyst = require('../assyst').client
+const assyst = require('./Assyst').client
 class Command {
     constructor(names, permissions, execute) {
         this.names = names;
